arreglos en el modulo de Empresa

+++ frontend/src/components/empresa/listaempresa.component.js
@@ -12,7 +12,7 @@ import Typography from '@material-ui/core/Typography';
 import EditIcon from '@material-ui/icons/Edit';
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
 import IconButton from '@material-ui/core/IconButton';
-
+import swal from 'sweetalert';
 
 import EmpresaService from '../../services/empresa.service'
 import Editarempresa from './editempresa.component'
@@ -35,11 +35,7 @@ export default class ListarEmpresa extends Component {
       empresas: [
         {
           id: '3333',
-          name: 'hfjegggghhghghghghfj',
-        },
-        {
-          id: '3555333',
-          name: 'hfjehfj',
+          name: 'ERROR',
         },
       ],
     };
@@ -68,15 +64,41 @@ export default class ListarEmpresa extends Component {
 
 
   deleteEmpresa(num) {
-    EmpresaService.delete(num)
-      .then(response => {
-        console.log(response.data);
-        //eliminado correctamente msj
-        //actualizar tabla
-      })
-      .catch(e => {
-        console.log(e);
-      });
+
+    swal({
+      title: "Esta seguro?",
+      text: "Una vez borrado no se puede recuperar!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+
+
+        EmpresaService.delete(num)
+        .then(response => {
+          console.log(response.data);
+          //eliminado correctamente msj
+          swal("Se ha borrado!", {
+            icon: "success",
+          });
+          //actualizar tabla
+          this.refreshList()
+        })
+        .catch(e => {
+          console.log(e);
+          swal("Error!", "no se logro borrar", "error");
+        });
+
+
+
+      } else {
+        swal("Cancelado!");
+      }
+    });
+
+    
   }
 
 
@@ -97,6 +119,8 @@ export default class ListarEmpresa extends Component {
           <Typography variant="h5" >
             Empresas
           </Typography>
+          
+          <br></br>
           <TableContainer component={Paper}>
             <Table style={{ minWidth: 650 }} aria-label="simple table">
               <TableHead>
@@ -124,6 +148,7 @@ export default class ListarEmpresa extends Component {
                           <Editarempresa
                             empresaid={row.id}
                             empresaname={row.name}
+                            refreshList={this.refreshList}
                           />
                         </Grid>
                         <Grid>
